+++ R/load.r
@@ -3,13 +3,17 @@ load = function (..., isolate = FALSE, renv = TRUE, dotenv = TRUE, dev = TRUE) {
 
   if (renv) {
     load_renv()
+    disable_renv_autoloader()
   }
+
   if (dotenv) {
     load_project_dotenv()
   }
+
   if (! isolate && ! load_user_rprofile()) {
     return(invisible(FALSE))
   }
+
   if (! isFALSE(dev)) {
     load_dev_package(dev)
   }